refactor: :recycle: Clase N°17: Tips para naming y abstracción de componentes React
