+++ index.html
@@ -41,7 +41,7 @@
     -->
 		<title>React App</title>
 	</head>
-	<body>
+	<body class="background-app">
 		<noscript>You need to enable JavaScript to run this app.</noscript>
 		<div id="root"></div>
 		<!--
