+++ src/components/ui/atoms/checkbox-icon/CheckboxIcon.tsx
@@ -1,12 +1,14 @@
 import { FC } from 'react';
 
+import styles from './CheckboxIcon.module.css';
+
 interface Props {
 	isCompleted: boolean;
 	updateIsCompleted: () => void;
 }
 
 export const CheckboxIcon: FC<Props> = ({ isCompleted, updateIsCompleted }) => (
-	<div className="a-checkbox">
+	<div className={styles['a-checkbox']}>
 		<input
 			type="checkbox"
 			defaultChecked={isCompleted}
