+++ src/main.tsx
@@ -3,7 +3,7 @@ import '@styles/index.css';
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { App } from './App';
+import { App } from './app/App';
 
 const container = document.getElementById('root') as HTMLDivElement;
 const root = createRoot(container);
