+++ src/utils/getLengthLeftTodo.ts
@@ -1,0 +1,4 @@
+import { Task } from '@src/types/interfaces';
+
+export const getLengthLeftTodo = (todos: Task[]): number =>
+	todos.filter((todo) => !todo.completed).length;
